refactor(venues): split SeatGridEditor under the line cap (#544)
Extract the config panel (SeatGridConfig) and the interactive grid +
its pointer-interaction cluster (SeatGrid) into sub-components; share
the SeatData type via seat-grid-types.ts. The SvelteSet/SvelteMap
selection model stays in the parent and is passed by reference so
cross-boundary reactivity is preserved. Behavior-preserving move only.
SeatGridEditor 853 -> 435 lines.
