+++ src/lib/components/venues/seat-grid-types.ts
@@ -1,0 +1,6 @@
+// Seat state stored per grid cell, keyed by "row-col".
+export interface SeatData {
+	exists: boolean;
+	is_accessible: boolean;
+	is_obstructed_view: boolean;
+}
